requirements/base.txt
pytz==2023.2  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.3  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower
uvicorn[standard]==0.21.1  # https://github.com/encode/uvicorn
qrcode
opencv-python-headless

# Django
# ------------------------------------------------------------------------------
django==4.0.10  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.53.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-jazzmin
django-filter==22.1
django-smart-selects
django-chartjs
djangorestframework-simplejwt
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.14.0  # https://github.com/adamchainz/django-cors-headers
drf-nested-routers
# DRF-spectacular for api documentation
drf-spectacular==0.26.1  # https://github.com/tfranzel/drf-spectacular




